+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.1.0 (January 4, 2013)
+
+* Added slide thumbnail options
+* Updated the default styling for slide indicators
+
 ### 1.0.0 (November 13, 2012)
 
 * Initial release